+++ src/whisper_streaming/online_asr.py
@@ -311,7 +311,7 @@ class VACOnlineASRProcessor(OnlineASRProcessor):
         import torch
 
         model, _ = torch.hub.load(repo_or_dir="snakers4/silero-vad", model="silero_vad")
-        from silero_vad_iterator import FixedVADIterator
+        from src.whisper_streaming.silero_vad_iterator import FixedVADIterator
 
         self.vac = FixedVADIterator(
             model
